FIX: LED intervall and radvalue stuck at 0

+++ arduino/nuclear-simulation/include/RadiationCalc.h
@@ -1,5 +1,5 @@
 
-void setupRadCalc(int,double,int);
-void updateRadCalcData(int,double,int); // (Reactor radiation per second, Room coefficient, protective gear coefficient)
+void setupRadCalc(int,float,int);
+void updateRadCalcData(int,float,int); // (Reactor radiation per second, Room coefficient, protective gear coefficient)
 float getRadCalcData();
 void updateRadTimer();
